Capture explictly to avoid unnecessary allocations
Summary:
Planning a convolution there was a disproportiante about of time spent constructing hte model (about half the time) compared to running the constraint solver.

This was because when we used `Model::call` we would use the capture-by-value lambda shortcut `[=]` instead of only capturing what we needed. This lead to an excessive amount of copies to objects that were in scope, expensive to copy and not required like the Target and ConvParams.

This change takes the number of allocations needed by the planner down by half (from 700k to 360k) and now 89% of the time is spent inside the solver and 9% constructing the model.

Reviewers: samuelh, richard, tomm, oleksiik

Reviewed By: tomm

Differential Revision: https://phabricator.sourcevertex.net/D17956