Reduce input range of the reduce_op test inputs for SQUARE_ADD, MULTIPLY
Summary:
There were numerical errors when increasing the number of random reduce tests used.  Some were checked that that
IPU/Sim gave the exact result expected if all arithmetic were done in half precision so there was no actual bug.
Limiting the range of numbers in the case of reduction that use a mpy should reduce the error in the output due
to the precision of the data.

Test Plan: Ran 60000 tests for reduce_op with random seeds.  No errors found.

Reviewers: samuelh

Reviewed By: samuelh

Differential Revision: https://phabricator.sourcevertex.net/D18272