Provide the factor for combineConvGroup in the transform
Summary:
Instead of calculating the factor from a bool, provide a range of factors and store them in
the transform struct. Enables the planner to select what it thinks is the best factor for the
convolution. If it cannot apply the transform reutrns 1 which should be a noop

Reviewers: samk, richard

Reviewed By: samk

Maniphest Tasks: T13180

Differential Revision: https://phabricator.sourcevertex.net/D17748