Apply constraints to poplin::ReduceAdd, use ld128 instruction, adjust benchmarks
Summary:
Apply the constraints, allowing use of ld128 and a speed up in poplin::ReduceAdd.  This improved various benchmarks by >1% so I updated these too, along with some that had a smaller improvement.

Note that the change to the cycle estimators can effect the decisions made by the convolution planner.  Consider also: the idea that convolution outputs to interleaved memory, therefore it's often ok if this reduction needs an input from interleaved memory.  That idea is OK, except if exchange happens to gather the data to reduce together- then this change means that the exchanged data now also needs to be in interleaved memory.

It is a matter of opinion if the speed improvement is worth the extra cost, I've tried to quantify this below.

Fixes T13380

Reviewers: samuelh

Reviewed By: samuelh

Maniphest Tasks: T13380

Differential Revision: https://phabricator.sourcevertex.net/D18541