Fix handling of POPFLOAT_ROUND_TO_ZERO in CastToGfloat32
Summary:
Previously it would round using POPFLOAT_ROUND_STOCHASTIC_EXTEND if
the rounding mode of the vertex was set to POPFLOAT_ROUND_TO_ZERO due to
a bug. We happened to get away with this in the tests since if the
PRNG is not initialised it always returns zero in which case stochastic
rounding gives the same results as round to zero.

Reviewers: Eddy, badreddine

Reviewed By: badreddine

Differential Revision: https://phabricator.sourcevertex.net/D18538