In convolution library rearrange weights before serial slice if necessary.
Summary:
Depends on D17777. Fixes T12469.

We just add a copy to rearrange weights if there is any serialisation of the convolution and leave it to poplar to remove it if it is unnecessary as it should be in most cases. This prevents unexpectedly laid out weights from producing terrible dynamic slice operations.

Reviewers: samk, richard

Reviewed By: samk

Maniphest Tasks: T12469

Differential Revision: https://phabricator.sourcevertex.net/D16393