Deal with cases where continuous reduce doesn't get targeted due to worker assignments
Summary:
Continuous reduce will frequently not get targeted when it should, as work was assigned in a round robin manner.  Therefore a worker would not be asked to reduce multiple sections of memory that were contiguous.  In this diff we avoid work splitting reductions with a suitable pattern length until we are about to create vertices.  Then we split them - but ask an individual worker to deal with a specific group of them which should target continuous reduce. Care is taken to pre-split groups of reductions where there are multiple regions - data from each region will become a new pattern and therefore a new group of reductions.

Fixes T13094

Reviewers: tomm, samuelh

Reviewed By: samuelh

Maniphest Tasks: T13094

Differential Revision: https://phabricator.sourcevertex.net/D18381