Allow calling GCL collectives
Summary:
Create and expose an interface for the ReplicatedCollectives such
that GCL can add their own implementation.

By default this does not change anything, the original poplibs
collectives are called as usual. But if the environment variable
GCL_NUM_IO_TILES is set calls to the replicated collectives are
redirected to the gcl implementation.

Reviewers: richard, emmanuel, samuelh, bartoszb

Reviewed By: richard

Subscribers: danielb

Differential Revision: https://phabricator.sourcevertex.net/D18560