Rearrange base tensor for multiSlice/multiUpdate before slicing/updating if it is poorly distributed in the introspected path.
Summary:
Fixes T13606.

Add a test which would fail without this facility as it is mapped over too few tiles (checked before/after and this is true).

I've opted to add the 'availableMemoryProportion' flag and rearrange if the max memory on a tile exceeds this proportion during the update. My reasoning is that this could be used during the planned slice/update later on and because it is hard to come up with a heuristic that does everything we might want without some kind of memory limit. This seemed simplest.

Reviewers: Eddy

Reviewed By: Eddy

Subscribers: matt, oskarb

Maniphest Tasks: T13606

Differential Revision: https://phabricator.sourcevertex.net/D17999