Improving no reduce / avoiding error when scale=1.0
Summary:
The reduction library copes with a no reduce case, where input=outputs. It sounds like there is nothing to do, but we could square the input, apply scale or do an update.

This would fail if we had update or squaring to do and scale was not required - it would access the scale tensor which didn't exist.  Additionally square and scale can be better done in float when the input is half which is more like the normal reductions do it.  Also now the map expressions cope with casting the casting can be absorbed into the expression which should become a fused operation.

Test Plan: Specific tests added to cover combinations of scale, half/float types, square and update.

Reviewers: samuelh

Reviewed By: samuelh

Differential Revision: https://phabricator.sourcevertex.net/D18918