Try and do expand dims transforms before allocating when the transformation is view-only.
Summary:
Fixes T11747. Ref T12224. Depends on D19088 (just for test coverage).
In an example case we have 1 input channel and 100 output channels. But we also have an input field dimension of size 200 and a matching kernel dimension of 200 which can be thought of as equivalent to input channels.

Reviewers: richard

Reviewed By: richard

Maniphest Tasks: T12224, T11747

Differential Revision: https://phabricator.sourcevertex.net/D19089