Implementation of 2 stage (on tile) reductions to solve 2 memory layout problems
Summary:
Address 2 problems, both of which are solved by implementing special cases of 2 stage reductions on a tile, based on detecting awkward memory layouts.

Where data is  to be reduced for columns A,B,C.. and the data is stored in memory on tile in a pattern: A,A,B,B,C,C,A,A,B,B,C,C ....The reduction can be carried out in 2 stages - the first just tolerates the layout and sums as if in columns:
AABBCCDDEEFF...
AABBCCDDEEFF...
.... x N rows
Reduces to :
AABBCCDDEEFF..

Then a second stage reduces to:
ABCDEF...

This solution can be applied for any innerFactor detected in reduction patterns - not just 2.  Splitting of reductions with innerFactor < 4 is implemented in this change, which seemed like a sensible threshold.

Division of work to try to use all 6 workers has been implemented in the 1st stage.  Work is divided by column, as dividing by row results in a very slow second stage.

------------------------------------------
The second problem is that small numbers of columns are awkward - notably 2 columns with data of type half.  The same issue could be observed for 3 columns, 5 etc - any number that doesn't work well with vector widths of 2(float) or 4(half).

For example, 2 columns, half data:
ABABABABABABAB..
Reduce with first stage as if we had 4 columns:
ABAB
ABAB
ABAB
... x N rows
Reduces to:
ABAB

The a second stage reduces to A, B

Division of work to try to use all 6 workers has been implemented in the 1st stage.  Work is divided by row - naturally we have very few columns.

-----------------------------------------
For both cases the main advantage was in avoiding the copies/inefficient reductions frequently arising from the memory layout.  Introducing 2 stages improved this dramatically.  Given that change, attempting to use all workers had a small additional improvement - depending on how easy it was to split work.
Note - temporarily absorbed the changes in groupNorm for a test case that this is meant to help with

Fixes T13651
Fixes T11221

Test Plan: Reliant on random reduction tests - increased these to 15000 and 1500 (multi ipu) locally to test more thoroughly

Reviewers: tomm, samuelh

Reviewed By: samuelh

Subscribers: ranjayc

Maniphest Tasks: T13651, T11221

Differential Revision: https://phabricator.sourcevertex.net/D17808