Simplistic fix to the interaction between sparse loss and reduction / small reduction enhancements
Summary:
Observed that using the specified test the sparse loss codelets (column variant) produced either 30 or 60 outputs per row (1024 rows).  The outputs are spread over multiple tiles. The resulting reductions are not all the same size (some 30, some 60) which results in creating multiple individual reductions.  With no output specified, these are all on tile zero.  So specifying the output and its mapping helps (Change specific to sparse loss).

However, those reductions with 60 outputs were attempted to be split into intermediate and final reductions.  Since the 2 constants used to plan this (Combined into one in this diff) appeared to be inconsistent, each reduction only had one intermediate reduction producing one intermediate partial for the final stage to deal with.  A waste of time, or at least confusing?!  This was only seen as the reduction size of 60 happened to be between the 2 constants 32, 64.  (Change to the reduction library)

If we did choose to split, Intermediate reductions would still have their outputs on tile 0, (as each call to reduce had only one row).  So, if we are only dealing with splitting a single reduction(we may have many single reductions), doing them all on tile 0 doesn't seem good so we pick up the tile number of the first partial and do the intermediate reduction on that tile .(Change to the reduction library)

Technically this fixes T10196 (already closed as the investigation was completed).  However a more carefully thought through approach to sparse loss would be better - see notes in T10196.

The changes to the reduction library in this diff are good general enhancements regardless of the sparse loss issue.

Reviewers: tomm

Reviewed By: tomm

Subscribers: O1 Reduction

Maniphest Tasks: T10196

Differential Revision: https://phabricator.sourcevertex.net/D13287