popops: Expose replicatedReduceScatter
Summary:
Expose replicatedReduceScatter() which is the first part of
replicatedAllReduce(). Verify assumption about rank-1 input.

We want to use this for hierarchical reductions in distributed
training. Then we can do replicatedReduceScatter() on the IPUs,
send the result from each replica to the host, and concatenate
the results there to form the per-worker reduced result. This
is then passed to an inter-worker allreduce routine to get the
final result.

This has some benefits compared to doing replicatedAllReduce()
and sending the result from one of the IPUs to the host:

1) replicatedReduceScatter() is less work than replicatedAllReduce().
2) Each replica only needs to transfer ceil(numElements/numReplicas)
   elements to the host (and they are done in parallel).

T14322

Test Plan: Added a new hardware test

Reviewers: richard, samuelh, stephenm

Reviewed By: richard

Subscribers: shauryas

Maniphest Tasks: T14322

Differential Revision: https://phabricator.sourcevertex.net/D19102