Elementwise fused ops with scalars
Summary:
Added the capability of fused operations to deal with expressions containing a mixture of tensors with >1 element (and all of the same shape) and tensors with 1 element (and any shape - eg {1} or {1,1,1} or {} etc.).  Additionally, where all tensors are have only 1 element, no vectorised code is generated. Vertex state for 1 element tensors consists of pointers only.

As per other fused ops this will only have any effect when the expression has more than one operation - so won't affect simple broadcasting operations at all - eg just adding a constant to a tensor.  It will only take effect when we have no complicated broadcasting to do (ie - we need all tensors to be the same shape but are now allowing scalars to be included) so it should not impact on cases where the non-fused binary ops need to identify aliases and pick a vertex based on patterns.

Reviewers: stephenm, samuelh, tomm

Reviewed By: samuelh

Differential Revision: https://phabricator.sourcevertex.net/D18931