Target wider innerFactor problem reductions with 2 stages.
Summary:
Change the effect of an observed case:
Partials: 0000111122223333....7777
              0000111122223333....7777
	      etc
So that it gets treated as 2 stages, the first giving 0000111122223333...7777
and then a continuous reduce gives: 01234567.

This avoids using the inefficient popops::Reduce<popops::ReduceAdd,half,float,false,1>

It is just an extension of the 2 stage approach to deal with problem column counts of 2,3 by just changing the controlling parameter!

Added some debug to this and other parts of reduction code.

This should improve the main issue in T14694

I could not find cases where this didn't result in an improvement.

Reviewers: #poplar_team, samuelh

Reviewed By: #poplar_team, samuelh

Differential Revision: https://phabricator.sourcevertex.net/D19794