Add an option to allow stable statistics calculation in norms
Summary:
The stable version requires the first order moment to be
estimated and applied to the sample set before the second
order central moment is calculated.

This is a straightforward implementation which uses all
the samples in the sample set. There exists stable algorithms
which estimate first order moment on a smaller sample set with
lower complexity. These could be a potential replacement for
this if real models show very large mean to standard deviation.

Fixes T15189

Reviewers: #poplar_team, tomm

Reviewed By: #poplar_team, tomm

Maniphest Tasks: T15189

Differential Revision: https://phabricator.sourcevertex.net/D20227