Cast isfinite's argument to float in generated codelets
Summary:
`isfinite` cannot be used with `half` type.
Generated codelets with `isfinite` calls in them also need to cast their arguments to `float` to avoid compilation errors.

This is normally done with `PromoteHalfsToFloats` in https://phabricator.sourcevertex.net/diffusion/POPLIBS/browse/master/include/popops/elementwiseCodelets.hpp$60

However, the codelet doesn't have access to this function (and if it needed it, handleSpecialCase would need to be modified to allow special cases to add includes to the top of the codelet.
`PromoteHalfsToFloats` essentially just performs a static_cast to float, so this is what I've done in the special case.

Please let me know if there's a more robust way of achieving this :)

Test Plan: Tested with a program that fails with the error when `isfinite` is called with a `half` - no longer throws the error and executes correctly.

Reviewers: stephenm, samk, #software_-_poplibs, #poplar_team

Reviewed By: samk, #software_-_poplibs, #poplar_team

Maniphest Tasks: T14636

Differential Revision: https://phabricator.sourcevertex.net/D19528