Add support for the SLIC 1x4 kernel to the convolution planner when the input type is half
Summary:
This was already landed as ecb10656e8df but was reverted as it contained a regression for existing plans.

This change is the old change plus the fix for the regression. Fixes T15152.

This reverts commit d03f355d980b3a7aeae3d438eb21e55c4c80bf9c.

Test Plan: tested manually, the layer that exploded has gone back to using the same plan as before.

Reviewers: tomm, richard, #poplar_team

Reviewed By: tomm, #poplar_team

Subscribers: oleksiik

Maniphest Tasks: T15152

Differential Revision: https://phabricator.sourcevertex.net/D20271