Fix memory corruption bug in the BroadcastClamp___half vertex
Summary:
Previously the code subtracted one from the loop count divided by two to
account for the fact that one iteration of the inner loop that handles
two elements has been peeled off. If the number of elements is 1 (or if
the number of elements is two but the base pointer is misaligned) this
subtraction underflows causing the vertex to write past the end of the
output vector. Fix this by checking the loop count before the subtraction.

Note the existing test case hits this issue, however we happen to get
away with it because there is nothing important stored in the memory that
is clobbered.

Fixes T15383

Reviewers: timp, oleksiik, #poplar_team

Reviewed By: timp, #poplar_team

Maniphest Tasks: T15383

Differential Revision: https://phabricator.sourcevertex.net/D20484