Rearrange 'slices' input to multiUpdateAdd under some circumstances to reduce exchange code/copy vertex state cost.
Summary:
Fixes T15023.

The slices that are the input to the multiUpdateAdd operation will typically be gradients from another layer in a network and so we don't have control over what the layout will be unlike the other operands to the update.

We check for a couple of different situations that we know will cause us increased memory use in terms of exchange code or copy vertex state if we need to rearrange the slices:
* We must cast up half to float partial type if we chose a two-stage multiUpdateAdd which is common when the slices are large (lots of updates). In this case we need to send/receive float partials rather than half partials i.e. we need to send/receive twice as many bytes.
* We must multi-cast the slices to multiple destination tiles. Here the cost of any rearrangement in terms of either exchange code or copy vertex state post-exchange is essentially multiplied by the factor by which we multi-cast the slices.

Reviewers: Eddy, #poplar_team

Reviewed By: Eddy, #poplar_team

Maniphest Tasks: T15023

Differential Revision: https://phabricator.sourcevertex.net/D20019