Removed tests (already commented out) in dynamicSlice codelet test files
Summary:
The tests that are commented out look to be no longer relevant.  They were introduced (commented out) in D12682, at that time it would have sounded reasonable to check that the change made worked (short became unsigned) for the numBaseElements parameter.  However this isn't generally done for most codelets.

But now -
The 2D vertex has been changed to use a vector list for baseT.  the code uses `baseT[r * numBaseElements].size()` to access it, with r=0..numRegions.  As the vector list cannot include anything like 2^16 -1 elements such a test would fail for that reason.

The supervisor case has vertex state:
```	  Input<unsigned> offset; // in \a baseT
	  Input<Vector<InType, ONE_PTR>> baseT;
	  Output<Vector<InType, ONE_PTR>> subT;
	  const unsigned numBaseElements;      // in the slice dimension
	  const unsigned short numSubElements; // in the slice dimension
	  const unsigned short regionSize;     // stride between slices```

In the code today there's nothing special about the treatment of `numBaseElements` - nothing that copes with its range being > 2^16 -1 it's not even used as a loop count.

Finally, the numbers suggested imply a large test being required which is usually slow (at least for copying initialised data in/out of the area of memory tested) - let's not introduce slow tests for no good reason.

Fixes T12986

Reviewers: #poplar_team, oleksiik, samuelh, Richardc, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: #poplar_team, oleksiik, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T12986

Differential Revision: https://phabricator.sourcevertex.net/D20848