Implicitly zero partials in half to float partials conv 1x1 vertex
Summary:
Replaces an existing implementation of the vertex which explicitly zeros
the partials in a separate loop. This should give modest savings when there
the product of the input and output channel groups is small. The cost paid
for the performance improvement is ~160 bytes of additional code size.

Fixes T15928

Test Plan: poplibs test (no new tests added)

Reviewers: tomm, #poplar_team, samk

Reviewed By: tomm

Subscribers: samk

Maniphest Tasks: T15928

Differential Revision: https://phabricator.sourcevertex.net/D20983