Apply more efficient  reduction vertices, avoiding specialisation 0,1 including multiple reductions consumed by one continuous reduce vertex
Summary:
Instead of taking a number of reductions and applying the inefficient specialisation0,1 vertex this diff allows for use of different specialisations.  For example, if a worker was assigned 6 reductions they can now be targeted (If the specialisation is suitable for implementing that reduction) using:
reduction [0,1) specialisation 2
reduction [1,5] continuous reduce
reduction [5,7) continuous reduce - but with partials of a different length
3 vertices for 6 reductions is still more efficient than using specialisation 0 or 1.

Also for the first reduction stage where on tile data is introspected, suitable patterns for targeting continuous reduce are not split up so this change will have no effect .  So this will only have an effect for intermediate and final reduction stages.

Fixes T12964 - Allows continuous reduce to target cases where the partials aren't continuous, but would be received onto a tile by exchange.
Ref T14694 - This is an extension of the method used to generate multiple vertices based on cost where we consider either single reduction  per vertex or all using 1 vertex.

In practice to do this the method of splitting reductions between workers needs to be changed - it is done in a round-robin manner, meaning that reductions with adjacent output elements will often be assigned to different workers, which stops us from being able to use continuous reduce without introducing copies.  This work is being continued in D21238

Reviewers: #poplar_team, samuelh

Reviewed By: #poplar_team, samuelh

Maniphest Tasks: T12964, T14694

Differential Revision: https://phabricator.sourcevertex.net/D20736