More variants of Reduction specialisation 3 in assembler
Summary:
Adding assembler variants of specialisation 3, allow for them to be targeted in ReductionConnection.cpp.

At the moment this will be targeted a little more due to covering the update option but won't do that much until we target vertices which cast float->half in the reduction library.  But when we do it should avoid some regressions in certain cases and improvements in others.

Fixes T16075

Test Plan: Added more variants of the ReduceSpecial23 testing

Reviewers: #poplar_team, samk, samuelh, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, Eddy

Reviewed By: #poplar_team, samk, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T16075

Differential Revision: https://phabricator.sourcevertex.net/D21270