Fix regression due to D20584
Summary:
D20584 caused one regression in the SDK tests (see T16937) but many improvements. Its effect was to create more reduction specialisation 3 vertices with a mixed type: float to half.  The regression is partly due to the library not picking these as the input width can now be 2 floats, giving an output of 2 halfs.  A more expensive specialisation 0 or 1 was picked instead.

Also noted that the change would result in >6 vertices per tile for the last reduction stage - so some workers have >1 vertex to run and become the slowest execution path.  This reverses the regression further.  Now we see a small improvement, in line with other benchmarks when D20584 was landed.

Fixes T16937

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Maniphest Tasks: T16937

Differential Revision: https://phabricator.sourcevertex.net/D22068