Cast before exchange to avoid costly copies when reductions require one half from any tile for inter/last stages
Summary:
If gathering a single half from any tile (usually multiple single halves from many tiles) in order to carry out an intermediate or last reduction stage, continuousReduce can be targeted but there is a costly copy step following exchange.  This can be avoided by casting to
float on tile before the the exchange happens.

That is just the same as the effect of an on tile reduction stage (for add, square add) which would produce float outputs - which would then be exchanged.

Unchanged in this diff: In the case of max, min reduce the process is more complex - we could cast to float before exchange but this is a different to what normally happens for max, min and further changes are needed.  Some of these are pending in D20584 so this diff only deals with add, square add reductions.

Fixes T15914

Test Plan: Verified by using the method described in T15914, for intermediate and last stages of reduction being the cause of the cast to float

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh, samk

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh, samk

Maniphest Tasks: T15914

Differential Revision: https://phabricator.sourcevertex.net/D21735