Added a conv group grouping to the internal shape of the convolution inputs
Summary:
This adds an extra level of grouping, for the conv groups, to the memory layout of the tensors used by the convolutions (input, weights and output).

It changes the shape like so:
 - activations (input/output): `[G][C/cpg][N]...[cpg]` to `[G/gpg][C/cpg][N]...[gpg][cpg]`
 - weights: `[G][Co/copg][Ci/cipg]...[copg][cipg]` to `[G/gpg][Co/copg][Ci/cipg]...[gpg][copg][cipg]`

The planner changed so that each method specifies a conv group grain size which the groups must be a multiple of (or else it gets padded). Currently only SLIC has a conv group grouping that isn't just 1.

This diff doesn't change anything, it is just a refactor to allow the new SLIC kernel.

Ref T15304. Depends on D22090.

Test Plan: added new unit tests, ran 20000 conv_random tests with zero failures. benchmarks unchanged as expected.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, godfrey.da.costa, tomm, richard, oleksiik, ranjayc

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, godfrey.da.costa

Maniphest Tasks: T15304

Differential Revision: https://phabricator.sourcevertex.net/D21055