Increase minimum size of outputs in intermediate reduction stages if minimum interval allows it
Summary:
The existing splitting strategy for intermediate reduction stages fixes
the splitting of reductions based on the average number of partials
used per tile in the full reduction. The splitting of the output is done
without accounting for the total tiles contributing to a partial output.

This change modifies the minimum outputs assigned to a tile thus increasing
the number of partials at the output of the stage. This in turn reduces
exchange code size as a larger chunk is exchanged.

Fixes T17223

Test Plan:
./test.sh poplibs
No regressions seen in benchmarks as this only targets cases when there are
a large number of tiles contibuting to an output

Reviewers: Eddy, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp

Reviewed By: Eddy, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T17223

Differential Revision: https://phabricator.sourcevertex.net/D22361