change tests and tools to avoid IPUModel unless they are in that mode
Summary:
ipumodel was used as a simple way to get some machine defaults (tilesPerIPU).
This isn't useful with the advent of mk2 which has a different number of tiles.

Now IPUModel is only used for IpuModel and IpuModel2 targets, the default is
not reported.

Depends on D22283, D22234
Fixes T17026

Test Plan: all poplibs tests continue to pass

Reviewers: saml, oleksiik, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, oleksiik, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T17026

Differential Revision: https://phabricator.sourcevertex.net/D22318