Attempt regroup flipped weights before copying to backward weights inweightsTransposeChansFlipXY
Summary:
Depending on the plan it is possible that operands may be swapped.
This results in the grouping not being very nice for transposition
as grouping is detected on ungrouped dimensions. In such cases,
there may still be a good grouping on the full tensor.

Fixes T17438

Test Plan:
Benchark test inception_tr_bs4_i9_c3x3 which had large rearrangement cost
does not show that rearrangement.

Reviewers: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: samk

Maniphest Tasks: T17438

Differential Revision: https://phabricator.sourcevertex.net/D22678