Explicitly copy convolution output to a more benignly mapped tensor if original layout is poor
Summary:
The convolution planner plans only for the layer and ignores how
its output layout affects operations which use the ouput or inherit
the layout in a model. If the layout is poor,  there is an adverse
effect on performance and memory over multiple operations.

This change defines a bad layout as one which doesn't have any
groupings and remaps the tensor to a more benign layout.

An option flag  to enable the remapping if detected is added. Only
single layer tests have this disabled  but the flag itself is enabled
by default.

Fixes T17486

Test Plan:
./test.sh poplibs
Add target test with option --remap-output-tensor=true

Reviewers: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Subscribers: samk

Maniphest Tasks: T17486

Differential Revision: https://phabricator.sourcevertex.net/D23015