Only remap output of convolution if output channels are a multiple of grain size
Summary:
Remapping with a fixed grain size was forced even if the
output channels per conv group, which is the dimension
that needs to be contiguous in memory, was not a multiple
of grain size. We could also look at other dimensions being
a multiple of the grain size, but this is not checked for
with this diff.

Fixes T17811

Test Plan:
Regressed TensorFlow tests are fixed
./test.sh poplibs

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T17811

Differential Revision: https://phabricator.sourcevertex.net/D23279