Move the IS_WORKER flag from the .stack_sizes section to the .text section header.
Summary:
This needs to be landed after D23533 and is a prerequisite for landing D19573 (main diff for T11207 - Stack size computation)

As defined in D23533, this moves the FORCE_IS_WORKER flag from the `DEF_STACK_USAGE` macro (bit 30 of the stack size value in the `.stack_sizes` section)  to the `.section .text.XXXX` directive  (i.e. uses bit 30 of the `sh_flags` field of the section header), for all poplibs functions that use that flag.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, javeda, Richardc

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, javeda, Richardc

Maniphest Tasks: T11207

Differential Revision: https://phabricator.sourcevertex.net/D23517