Remove the use of rebalance tensor in dynamic slice
Summary:
Even when createSliceableTensor is used, dynamic slice doesn't
return a tensor which has the same mapping as how the sliceable
tensor was created. This is because tensor is rebalanced when
dynamic slice is called. The rebalancing is done in a way that
doesn't match groupings even if they exist and results in
poor copies. With features being added to create sliceable
tensors, it should be the callers responsibility to balance
the tensor slices as the dynamic slice should retain the same
mapping as used in the update.

Fixes T18270

Test Plan: pipelined resnet and lstm SDK tests show upto ~9% improvement.

Reviewers: tomm, Eddy, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T18270

Differential Revision: https://phabricator.sourcevertex.net/D23626