Map output of reductions and parameters in norms based on their use
Summary:
Biases, norm parameters(beta and gamma) and reduction
outputs are typically broadcast and passed as one of the
operands in an arithmetic operation.

This change maps the tensor based on the use of the
the activations/gradient tensor on tiles. Elements are
only mapped on tiles where the gradients/activations are
used.

Fixes T16321

Test Plan: Added nre test

Reviewers: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: lakshmik, oskarb

Maniphest Tasks: T16321

Differential Revision: https://phabricator.sourcevertex.net/D21357