Updated SLIC tool to call the actual method in the convolution library
Summary:
Ref T18279.

As the tool was written first this logic ended up getting duplicated in the library. To make it easier to extend the SLIC vertex this change removes the duplication. The tool now calls the same method to create the vertex that the library does.

Test Plan: existing tests still pass

Reviewers: tomm, claudiop, timp, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T18279

Differential Revision: https://phabricator.sourcevertex.net/D23745