Strided reduce as a replacement for specialisation 3
Summary:
The idea is to replace/modify specialisation 3 to be a strided reduce vertex by adding a partialsWidth parameter which allows the whole partials region to be left in one place in memory, but have different workers deal with a part of it.  This potentially replaces specialisation 3 and reducePartialsEqualSize.  This makes it a candidate to be a supervisor vertex but the idea of a supervisor vertices for reduction is maybe not a great one, so at the moment it's a worker vertex.

Pros and cons compared to the vertices it's replacing as a  worker vertex in this change:
Vertex comparison
Specialisation3: The inner loop is just as fast as this but we have 2 bytes more vertex state for partialsWidth.  If needs be we could get the best of both worlds by implementing both and the assembler would just have 2 different entry points, or maybe pack the information into 16 bits!  Otherwise it's identical.
partialsEqualSize:  We can support a grainSize of 4 halfs in the column count - so work can be divided better than with the value of 8 for partialsEqualSize.  Vertex state is smaller (both fewer bytes in immediate pointers and no DELTAN needed) .  The inner loop is faster: 1 cycle for 64 bits inputs vs 4 cycles for 128 bits.  The only downside is that it can only be applied with no copies being introduced when the layout in memory is regular (can be described by a stride and in one region) - the flexibility of the "vector list" vs "pointer, stride".

Observed: Improvements in most of the reduce benchmarks in poplibs, no regressions.  Simple tests improve as we avoid copies (gathering partials so that we can use the old specialisation3) and better work splitting in the 1st stage due to the possibility of grain size 4 halfs, 2 floats.

The way it's targeted is:
1. When patterns are analysed in the 1st on tile reduce we detect a simple regular 1 region pattern and write allPartials, offset, stride.  Where we have allPartials and the shape is good (4 halfs partials etc...) we then use stridedReduce "properly".
2.  As a replacement for the old specialisation3 when allPartials isn't initialised. This can happen in any stage but tends to be in intermediate and last stages.  The striding ability isn't used here!
3. When gathering partials together for the final reduction stage we can form them all into one partial and add this layout information if that would be a regular shaped region and works for all reductions.

Ref T14526

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samuelh

Subscribers: tomm

Maniphest Tasks: T14526

Differential Revision: https://phabricator.sourcevertex.net/D22172