Zero output padding after the convolution rather than before when using SLIC.
Summary:
Fixes T18543.

We do this because the zeroed padding was being trampled due to the fact that it was connected to an Output edge in the SLIC vertex and hence could be overwritten. One fix is to make that an InOut edge to track the previous values, but this means the output is live during the exchange to convolve. We can just perform the zeroing of output padding after the convolution instead.

Reviewers: samk, timp, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: samk, timp, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T18543

Differential Revision: https://phabricator.sourcevertex.net/D24318