T18732: combine same sized convolutions
Summary: Extra optimisation passes can be added to the multi-convolution API before calling through to the actual implementation. An obvious one (e.g. When the alpha ratio is 0.5) is when more than one of the convolutions passed in are the same shape, we combine these in the group dimension first.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, saml

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk, saml

Maniphest Tasks: T18732

Differential Revision: https://phabricator.sourcevertex.net/D24617