SLIC stride !=1 allowed in all output dimensions
Summary:
Remove constraints on strides in all but the innermost dimension (this was already allowed but constrained to 1 or 2) and remove constraints on output truncation.  Add inputs to test these to the tool.  Appears to function, but the resulting plans are sometimes faster, sometimes slower - see below. This cause of this was addressed in D24678, resulting in all affected benchmarks now being faster.

Fixes T18282

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, samk

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Maniphest Tasks: T18282

Differential Revision: https://phabricator.sourcevertex.net/D24494